+++ site/next.config.mjs
@@ -1,8 +1,13 @@
+import { createMDX } from "fumadocs-mdx/next";
+
+const withMDX = createMDX();
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   output: "export",
   images: { unoptimized: true },
+  reactStrictMode: true,
   trailingSlash: true,
 };
 
-export default nextConfig;
+export default withMDX(nextConfig);